+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- WAExHelper to build WebApplicationException
+
 ### Changed
 
 - native-helper-maven-plugin 1.3.5